backend/functions/sms/requirements.txt
twilio==6.5.0
firebase-admin
google-cloud-firestore
mock
attrs
phonenumbers
names
sentry-sdk==0.14.3